Issue #15493: Undpeloy synchronously in the same event listener thread in order to avoid deadlock observed when we used to hand off the undeployment task to the separate deployer/undeployer thread.
If there is a current deployment task pending, we now wait for it to finish. The deployment task
aborts if it detects at the beginning of execution that the bundle is stopping.

Deployment now returns a non-null object to indicate that really it got deployed. This non-null value
is used by undeployment task as well.

See JIRA for change control process details.

svn path=/trunk/; revision=44377


Former-commit-id: 286d7f57913c74975fd432e51907600b2e970a30