Fix for http://java.net/jira/browse/GLASSFISH-15576 "admin gui upgrade failure when grizzly schema upgrade doesn't happen first"
This change injects an instance of GrizzlyConfigSchemaMigrator into AdminRESTConfigUpgrade to ensure that the former is run first. Otherwise the upgrade can fail depending on the order that the modules are run. This is a problem we've seen in the promoted builds but not in the trunk of the workspace. For instance, upgrades work without this fix in the trunk, but fail in https://svn.java.net/svn/glassfish~svn/tags/3.1-b37

Change approved by Chris.
Code reviewed by Jerome.


svn path=/trunk/; revision=44560


Former-commit-id: 0828eeb4335c7e138fdba7a3d51fc89656061f73