Fix for 15624
If an app contained a file named using a special character - one that would be encoded in URLs - then the payload implementation failed during synchronization as it tried to expand the payload uploaded to the instance into the instance's directories.  

This change repairs that problem.

Approved: Tom
Code review: Tom
Tests: error scenario in the issue, QL tests, deployment devtests with upload

svn path=/trunk/; revision=44653


Former-commit-id: bb61f23e36a7fa31c392251aae0a88b3647d2325