Fix for GLASSFISH-15660
EJBs bundled within the bundled library in the WAR (jar in WEB-INF/lib) was represented as EJBs of the WAR BDA to Weld. The EJBs were represented as Bean classes in the WEB-INF/lib jar BDA. So, Weld understood the EJB as both a ManagedBean and an EJB, and hence deployment failed with an Ambiguous Resolution exception. 

Fix involved ensuring that the WEB-INF/lib jar BDA has the EJB Class represented as EJBs and the Bean classes of that BDA.

Tests run: QL(Classic and Web profile), CDI dev tests with and without security manager, CDI TCK, test case provided by bug submitter
Approved by: Chris Kasso


svn path=/trunk/; revision=44711


Former-commit-id: a49f99967993fc2f83e4d3a33645c3f3e7fd81ea