Fix for 15716
appclient -help displays the usage output, not the much more complete help output.  This is because the code checks to make sure the user specifies a main class or a client JAR; if not, it automatically acts as if the user entered "-usage" instead.  But that logic failed to account for cases in which the user specified -help (or -usage for that matter).  And because the ACC checked for "-usage" before "-help" it would display the usage text instead.

These changes fix that.

Approved: Chris
Reviewed: Hong
Tests: updated appclient devtests

svn path=/trunk/; revision=44748


Former-commit-id: 594a9a572128911686aa4227f58e4e5dd25389c3