Fix for 13774.
These changes implement a "stale" file handling mechanism for the FileArchive class which prevents stale files that were left over from a previous use of the same directory from being visible to a client of the FileArchive.  This scenario would happen with some frequency on Windows when an app was deployed with a name, a JAR file remained open and so the expanded directory in applications/ could not be deleted, and a different app entirely was deployed using the same name.  The applications/(appname) directory would be reused - since it remained from the previous app - and the stale file(s) within would confuse GlassFish.  Now, such stale files are detected and are suppressed so they are not treated as part of the FileArchive.

A hidden file at the top level of the FileArchive directory records such stale files so even after a server restart the stale files are not visible.

pom change approved by Sahoo and Jane
Test: deployment devtests on Mac OS X and Windows; error scenario on both OS types; added unit tests

svn path=/trunk/; revision=45495


Former-commit-id: 63dbfe69311fe9756ff870eab7ce8a09f0fdb24c