http://java.net/jira/browse/GLASSFISH-16197
Before Scenario:
1. Server is started with --passwordfile argument
2. the password file is deleted or renamed or made unreadable
3. asadmin restart-[instance|domain]

result -- the server is stopped but not started again because the password file is gone.  Since the command is asynchronous we
could not return an error message,  User waits 10 minutes for the timeout.  server.log says nothing.

After Scenario:

restart commands now first chack and make sure that the password file exists and is readable (if applicable)
If not -- the restart commands fails, the server continues runningm and the user gets an error message.





svn path=/trunk/; revision=45515


Former-commit-id: 4ce0c3eb9ebeb703ef7424807a941fc12835cf9f