issue 15987 - make a change to a system-property that is referenced by JVM option cause restart-required notification
This change makes SystemPropertyListener and GenericJavaConfigListener independent services that are automatically
started by AppServerStartup rather than having AdminAdapter start them.  Also, these services listen on the object
that they are interested in automatically.  The Cluster is not injected into the habitat with the DEFAULT_INSTANCE_NAME
key so that it can be injected directly into an object. 

The GenericJavaConfigListener has been modified to listen for system-property changes in the domain, server, cluster, and config objects and if a property referenced by the JavaConfig changes, then a restart-required notification is triggered.

Passed QL.

svn path=/trunk/; revision=45608


Former-commit-id: bb73be2d30f2c6d77da36bd306a838b743c89e56