Fix for 16261
The @PreDestroy method on an app client class (whether specified using annotations or in the descriptor) was not called.

The clean-up logic in the ACC was correctly written to invoke this but the injection manager in the clean-up objects was never set to the injection manager for the main class, so the clean-up logic could not invoke the pre-destroy method in the app client.

These changes store the injection manager with the clean-up information so the clean-up logic can invoke the method.

Tests: QL, deployment dev tests

svn path=/trunk/; revision=45720


Former-commit-id: 8658959f6b7865324d954924cc6d042766a1597d