Incremental fix for 16300
These changes suppress some incorrect error messages; pre-existing directories cause File.mkdirs to return false which the PayloadFilesManager was incorrectly reporting as a more serious problem.  The change to AsadminMain is to bring the CLILoggerHandler into compliance with the Logger.log contract (it was not substituting arguments for placeholders in the log string).

Tests: QL

svn path=/trunk/; revision=45983


Former-commit-id: b6811249617028844fbf4fd856c3eebeed98632c