Added FileUtils.mkdirsmaybe(File f)
It returns true if and only if the directory exists at the end of the call
It returns false if and only if the directory doesn't exist and can't be created.

I.e. it is handy for when you want to create a directory if it isn't already there - but you don't care if it already exists.



svn path=/trunk/; revision=46248


Former-commit-id: e6b0ffe5cce79b7b7e45fba020114c08195544ea