"Fix" for 16444
This is actually a workaround for an apparent limitation in IBM's Java on AIX.  The options string 

-javaagent:pathToAgent=options

seems to be truncated if the entire -javaagent token exceeds 512 characters (give or take a couple of characters).  

Some of the ACC work is done by a Java agent so it can get control before the user's client (and so the gf-client.jar can be on the classpath without requiring it to be explicitly specified in -classpath).  The options value passed to the agent can be quite large because it includes most of the command-line arguments and options the user specifies when invoking the appclient script - and then some.  The large value combined with a potentially long path to the the agent JAR caused the options string which Java passes to the agent to be truncated.

This fix causes CLIBootstrap to write the options to a file, then specify the file in the options string.  Then the app client agent reads the actual options from the temp file before deleting it (unless a private, unpublished property is set to keep the options file for diagnosing any problems that might occur).

Approved: Sathyan
Tests: QL, deployment devtests, spot tests on AIX

svn path=/branches/3.1.1/; revision=46488


Former-commit-id: 200b0efce6a63a14d6aeb5f196443a2150045bec