Fix for Bug 12408108 - WINDOWS XP ONLY: SIGTEST DEPLOYMENT ERROR WHEN DEPLOYING SIGTEST EAR FILE
During a Windows deployment, a JAR in WEB-INF/lib was opened but never closed.  When the app was undeployed Windows could not remove the file.  

During a subsequent redeployment the left-over library JAR is incorrectly detected as a submodule archive because the ApplicationArchivist uses File directory scans rather than the FileArchive API.  

These changes (1) make a factory method in FileArchive for the StaleFileManager public, and (2) enhance ApplicationArchivist to use the StaleFileManager to filter what files are visible.  This correctly hides the left-over library JAR from being discovered and misinterpreted as a submodule archive.

svn path=/trunk/; revision=46541


Former-commit-id: 0eefa22fcf9ad2897f77f740dfcb9f7f078f7a05