Submitted by: Dies Koper Reviewed by:  Naman Mehta
I've seen a lot of cases where two calls to a logger API follow like that:

  _logger.log(Level.WARNING, "ejb.system_exception",
              new Object[]{ejbDescriptor.getName(), inv.beanMethod});
  _logger.log(Level.WARNING, "", inv.exception);

Basically, what the developer wanted logged was:

[#|10T16:56:22.378+1000|WARNING|glassfish3.2|logger|ThreadID|EJB5184:A system exception occurred during an invocation on EJB HelloStateful, method: public void com.sun.mdb.HelloStateful.hello(java.lang.String)
javax.ejb.NoSuchObjectLocalException: The EJB does not exist. session-key: 490bfe700a81f-ffffffffd8b0d24f-1
	at
[...]
|#]

I.e. a message followed by e.printStacktrace(), but there is no Logger API for that!
But currently, the "" (or null) message argument leads to a "The log message is null." as the log component expects a proper message (key).

This change makes this possible on the logger side, so the code above would log the following:

[#|10T16:56:22.378+1000|WARNING|glassfish3.2|logger|ThreadID|EJB5184:A system exception occurred during an invocation on EJB HelloStateful, method: public void com.sun.mdb.HelloStateful.hello(java.lang.String)
javax.ejb.NoSuchObjectLocalException: The EJB does not exist. session-key: 490bfe700a81f-ffffffffd8b0d24f-1|#]
[#|10T16:56:22.378+1000|WARNING|glassfish3.2|logger|ThreadID|javax.ejb.NoSuchObjectLocalException: The EJB does not exist. session-key: 490bfe700a81f-ffffffffd8b0d24f-1
	at
[...]
|#]

When no exception and no exception is passed (which are blatant bugs), an error message is logged. I've clarified in the message to log a bug against the component mentioned in the logger field (as opposed to the logger component).



svn path=/trunk/; revision=46850


Former-commit-id: 54351799a826d2071c2a51c3e531fe9ec8a01f91