JIRA 16725
Problem:  If the server is started without the -javaagent JVM option then ProbeProviderClassFileTransformer.java gives up 
right away and monitoring will not work because the agent class can't be loaded.

Solution:
Instead of giving up -- try to attach to the agent.  Once that is done the class can be loaded and monitoring will work.




svn path=/trunk/; revision=47110


Former-commit-id: 9131acfaef8c03e47696b79a44c5e8001c23f8ab