JIRA 16725
Problem:  If the server is started without the -javaagent JVM option then ProbeProviderClassFileTransformer.java gives up 
right away and monitoring will not work because the agent class can't be loaded.

Solution:
Instead of giving up -- try to attach to the agent.  Once that is done the class can be loaded and monitoring will work.




svn path=/branches/3.1.1/; revision=47111


Former-commit-id: 74f194698f5250049a771841e7cf46aace66b59f