Fix for 16719
When a new instance is created the domain.xml from the DAS was included in the bootstrapped files.  This seemed to confuse things if the new instance was started when the DAS was down, thus preventing a sync.

These changes cause the bootstrapping to no longer include domain.xml.

Approved for 3.1.1: Sathyan
Tests: admin CLI cluster tests

svn path=/branches/3.1.1/; revision=47152


Former-commit-id: 533df7f9ad8568fcd7dd294b27415a6e9e2fb93c