16762
MiniXmlParser had no idea that the enabled attribute for profiler has a default value of true.  
A unit test immediately broke when I changed it.  Which is nice.  The unit test was also changed to reflect 
the change.

I used the excellent suggestion that Tom added to the issue.  Thanks, Tom!


svn path=/branches/3.1.1/; revision=47231


Former-commit-id: 2e6f9b125fc3f19214806f2940e57f723edcc009