JIRA 16287
This is where the process utilities I've added pay off.  No more fooling around.  To do a restart we now
wait for the OS itself to declare that the old server's process is really really dead and buried.
Before we did half checks like seeing if the admin port went deaf. Or that the standard streams closed. 
That turned out to be not good enough.


svn path=/trunk/; revision=47314


Former-commit-id: c2e116f0639d40a1378f3d42523e0a313c5620d2