JIRA 16801
ThreadPoolProbeProvider fixes from an earlier checkin did not quite do the job for 3.1.1
The calling classes (Grizzly) are differently organized in 3.1.1 vs. 3.2 which is why it was missed.

What happened is that the probe itself has String as a param.  The code that calls the probe sends in a String.  But
the listener -- which the probe calls -- is expecting a long.
All three have to match.  I fixed the other 2 to use long instead of String


Thanks to Sahoo for finding this.


svn path=/branches/3.1.1/; revision=47337


Former-commit-id: ba99fec72d7e7dce6775eb7c82182d11e8578b06