Fix for 16842
During deployment, GlassFish traverses the transitive closure of an app client JAR and all JARs it depends on.  This is how GlassFish discovers the full set of JARs that need to be downloaded as part of "asadmin deploy --retrieve" or "asadmin get-client-stubs" or a Java Web Start launch of the client.

The logic in app client deployment to detect (and traverse only once) a dependency cycles among JARs (... JAR A depends on JAR B depends on ... depends on JAR A) was incomplete.  It would just loop, following the cycle repeatedly until it exhausted help space.

These changes complete the cycle detection and avoidance checking.

Tests: test app from customer

svn path=/trunk/; revision=47424


Former-commit-id: d8398916431c47c0c899d6e5d05fdec557611ba7