Fix for 16130
When the AdminAdapter rejects a request from a remote node because remote admin is not enabled, it sets only the status in the response - it does not also set the response message.  On the asadmin client side, if the http status returned is not OK then that logic essentially assumes there is nothing of value in the payload - such as the action report - and displays only the response message.

These changes cause the AdminAdapter to set not only the response status to Forbidden but also to set the response message to the more user-friendly text so the asadmin client as-is will display it.

Approved for 3.1.1: Sathyan
Tests: QL

svn path=/trunk/; revision=47479


Former-commit-id: dd03eda4cd9355cbd7f33e600254db2248a8b9cc