Fix for 17040
The ACCAppClientArchivist is used only in the ACC to read client archives.  But because it is an Archivist, the deployment infrastructure in the DAS loads this class, which causes the gf-client-module module to be activated during any deployment, regardless of whether the app contains an app client.

These changes move the ACCAppClientArchivist out of gf-client-module into the deployment/dol module.  As a result gf-client-module is not unnecessarily activated. 

Approved for 3.1.1: Sathyan
Tests: QL, deployment devtests

svn path=/branches/3.1.1/; revision=48031


Former-commit-id: bf077bceba7dbf112c95e7f33d2ca935809ab6d8