Add support for AfterReplication supplemental commands.
Supplemental commands annotated with AfterReplication (as opposed to Before or After) will be executed after the command framework has replicated the command to any instances required.  Such commands will be run even if no clusters or instances exist, or if a target cluster exists but has no instances in it.

Tests: QL

svn path=/trunk/; revision=49039


Former-commit-id: be4d90834629b6623bc86978dbc58ec7bc4ff03d