Fix for 17319
Life cycle modules appear in the configuration as applications (as do more usual apps) but they do not have locations - directories - associated with them.  Some of the versioning code assumed that all apps have a location.

This change fixes such a problem exposed when an app and a life cycle module were both present and the user dynamically reloaded the app (using "touch .reload").

svn path=/trunk/; revision=50094


Former-commit-id: dec6cbbee550144ffdc975f59aa6171cae897a9a