Fix for 17371
It turns out to be hard to correctly accept and handle a string that can be either a file path or a URI.  Because AsadminInput is used only internally I'm changing the behavior to treat the string as either a dash (for reading from stdin) or a file path.  Consulted with Byron about the change and the resulting changes to NodeRunnerDcom.

Tests: QL, deployment devtests

svn path=/branches/3.1.2/; revision=50886


Former-commit-id: 58aa581891aead5e16978cfbd0a69ac25dbe736d