Fix for 17371
It turns out to be hard to correctly accept and handle a string that can be either a file path or a URI.  Because AsadminInput is used only internally I'm changing the behavior to treat the string as either a dash (for reading from stdin) or a file path.  Consulted with Byron about the change and the resulting changes to NodeRunnerDcom.

Tests: QL, deployment devtests

svn path=/trunk/; revision=50887


Former-commit-id: ad0eca6039f669447224a26808cb50ad017b58ce